snowflakestructuredpropertydemo/requirements.txt
requests>=2.28.0
snowflake-connector-python>=3.0.0
datahub>=0.12.0
python-dotenv>=1.0.0
